Rename stdlib to lib
