+++ ballerina/init.bal
@@ -21,5 +21,5 @@ isolated function init() {
 }
 
 isolated function setModule() = @java:Method {
-    'class: "io.ballerina.stdlib.data.xmldata.utils.ModuleUtils"
+    'class: "io.ballerina.lib.data.xmldata.utils.ModuleUtils"
 } external;
