+++ build-config/resources/CompilerPlugin.toml
@@ -1,6 +1,6 @@
 [plugin]
 id = "constraint-compiler-plugin"
-class = "io.ballerina.stdlib.data.xmldata.compiler.XmldataCompilerPlugin"
+class = "io.ballerina.lib.data.xmldata.compiler.XmldataCompilerPlugin"
 
 [[dependency]]
 path = "../compiler-plugin/build/libs/data.xmldata-compiler-plugin-@project.version@.jar"
