+++ compiler-plugin-test/src/test/resources/testng.xml
@@ -21,7 +21,7 @@
 <suite name="BallerinaConstraintCompilerPluginTests">
     <test name="UnitTests">
         <classes>
-            <class name="io.ballerina.stdlib.data.xmldata.compiler.CompilerPluginTest"/>
+            <class name="io.ballerina.lib.data.xmldata.compiler.CompilerPluginTest"/>
         </classes>
     </test>
 </suite>
