+++ compiler-plugin/src/main/java/io/ballerina/lib/data/xmldata/compiler/XmldataCodeAnalyzer.java
@@ -16,11 +16,24 @@
  * under the License.
  */
 
-module io.ballerina.stdlib.data {
-    requires io.ballerina.runtime;
-    requires io.ballerina.lang.value;
-    requires java.xml;
-    requires junit;
-    requires org.apache.commons.lang3;
-    exports io.ballerina.lib.data.xmldata.xml;
+package io.ballerina.lib.data.xmldata.compiler;
+
+import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.projects.plugins.CodeAnalysisContext;
+import io.ballerina.projects.plugins.CodeAnalyzer;
+
+import java.util.List;
+
+/**
+ * Xmldata Code Analyzer.
+ *
+ * @since 0.1.0
+ */
+public class XmldataCodeAnalyzer extends CodeAnalyzer {
+
+    @Override
+    public void init(CodeAnalysisContext codeAnalysisContext) {
+        codeAnalysisContext.addSyntaxNodeAnalysisTask(new XmldataRecordFieldValidator(),
+                List.of(SyntaxKind.MODULE_PART));
+    }
 }
