--- compiler-plugin/src/main/java/io/ballerina/stdlib/data/xmldata/compiler/Constants.java
@@ -1,17 +1,0 @@
-package io.ballerina.stdlib.data.xmldata.compiler;
-
-/**
- * Constants for XmlData's compiler plugin.
- *
- * @since 0.1.0
- */
-public class Constants {
-    static final String PARSE_STRING = "parseString";
-    static final String PARSE_BYTES = "parseBytes";
-    static final String PARSE_STREAM = "parseStream";
-    static final String PARSE_AS_TYPE = "parseAsType";
-    static final String TO_XML = "toXml";
-    static final String NAME = "Name";
-    static final String NAMESPACE = "Namespace";
-    static final String XMLDATA = "xmldata";
-}
