+++ native/src/main/java/io/ballerina/lib/data/xmldata/utils/ModuleUtils.java
@@ -16,11 +16,31 @@
  * under the License.
  */
 
-module io.ballerina.stdlib.data {
-    requires io.ballerina.runtime;
-    requires io.ballerina.lang.value;
-    requires java.xml;
-    requires junit;
-    requires org.apache.commons.lang3;
-    exports io.ballerina.lib.data.xmldata.xml;
+package io.ballerina.lib.data.xmldata.utils;
+
+import io.ballerina.runtime.api.Environment;
+import io.ballerina.runtime.api.Module;
+
+/**
+ * This class will hold module related utility functions.
+ *
+ * @since 0.1.0
+ */
+public class ModuleUtils {
+
+    /**
+     * Time standard library package ID.
+     */
+    private static Module module = null;
+
+    private ModuleUtils() {
+    }
+
+    public static void setModule(Environment env) {
+        module = env.getCurrentModule();
+    }
+
+    public static Module getModule() {
+        return module;
+    }
 }
